feat: add option `exceptNotNullishAlternates`
To except not nullish alternate conditionals from being reported in
`preferAndOperator` mode.
Some may want to just convert to AND operators if the alternate is
nullish as it may look cleaner. For people that may want to do that
but except cases where the alternate is not nullish and prefer a ternary
over two conditional expressions